feat: refactor layout components and add phone number formatting utility
