+++ src/entities/website/Footer.tsx
@@ -1,20 +1,14 @@
 import React from 'react';
 import { Container, Group } from '@mantine/core';
 
-type FooterProps = {
-    left?: React.ReactNode;
-    center?: React.ReactNode;
-    right?: React.ReactNode;
-};
-
-export const Footer = ({ left, center, right }: FooterProps) => {
+export const Footer = () => {
     return (
         <footer>
             <Container size="xl" py="md">
                 <Group justify="space-between" align="flex-start">
-                    <div>{left}</div>
-                    <div>{center}</div>
-                    <div>{right}</div>
+                    <div>left</div>
+                    <div>center</div>
+                    <div>right</div>
                 </Group>
             </Container>
         </footer>
